+++ restaurantbooking/src/main/java/com/project/restaurantbooking/JadeConfiguration.java
@@ -53,14 +53,14 @@ public class JadeConfiguration {
             restaurantAgent.start();
 
             // Initialize the GatewayAgent
-            JadeGateway.init("com.project.restaurantbooking.agent.TheGatewayAgent", jadeProfile.getBootProperties());
-            //Start Gateway Agent
-//            JadeGateway.init("com.project.restaurantbooking.agent.RestaurantGatewayAgent", jadeProfile.getBootProperties());
+//            JadeGateway.init("com.project.restaurantbooking.agent.TheGatewayAgent", jadeProfile.getBootProperties());
+
 
         } catch (StaleProxyException e) {
             e.printStackTrace();
         }
-
+        //Start Gateway Agent
+        JadeGateway.init("com.project.restaurantbooking.agent.RestaurantGatewayAgent", jadeProfile.getBootProperties());
         return jadeContainer;
     }
