+++ restaurantbooking/src/main/java/com/project/restaurantbooking/agent/RestaurantGatewayAgent.java
@@ -1,20 +1,19 @@
 package com.project.restaurantbooking.agent;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.project.restaurantbooking.messagetemplates.AddStaffRequest;
-import com.project.restaurantbooking.messagetemplates.ChangeStaffRequest;
-import com.project.restaurantbooking.messagetemplates.DeleteStaffRequest;
-import com.project.restaurantbooking.messagetemplates.LoginRequest;
+import com.project.restaurantbooking.messagetemplates.*;
 import jade.core.AID;
+import jade.core.behaviours.CyclicBehaviour;
 import jade.lang.acl.ACLMessage;
 import jade.wrapper.gateway.GatewayAgent;
+import lombok.SneakyThrows;
+import org.json.JSONObject;
+
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 public class RestaurantGatewayAgent extends GatewayAgent {
-
-//    protected void setup(){
-    //super(setup())
-//        System.out.println("Restaurant agent started.");
-//    }
+    private Map<String, AgentCommand> pendingCommands = new ConcurrentHashMap<>();
 
     @Override
     protected void processCommand(Object command){
@@ -86,8 +85,75 @@ public class RestaurantGatewayAgent extends GatewayAgent {
             send(changeStaffRequestMessage);
             this.releaseCommand(changeStaffRequest);
         }
-        else{
+
+        if (command instanceof AgentCommand) {
+            System.out.println("\nGatewayAgent - command: " + command);
+            AgentCommand agentCommand = (AgentCommand) command;
+            pendingCommands.put(agentCommand.getCorrelationId(), agentCommand);
+
+            ACLMessage msg = new ACLMessage(ACLMessage.INFORM);
+            msg.addReceiver(new AID(agentCommand.getTargetAgent(), AID.ISLOCALNAME));
+            String agentCommandJSON;
+
+            try {
+                agentCommandJSON = agentCommand.getContent(); //objectMapper.writeValueAsString(agentCommand);
+                msg.setConversationId("reserve");
+                msg.setProtocol("reserve");
+                msg.setContent(agentCommandJSON);
+                send(msg);
+                System.out.println("\nGateway: Msg Sent to RestAgent\n");
+            } catch (Exception e) {
+                System.out.println("Error: " + e.getClass().getSimpleName() + " - " + e.getMessage());
+                agentCommand.completeFutureResult("Error: " + e.getMessage());
+                pendingCommands.remove(agentCommand.getCorrelationId());
+                this.releaseCommand(command);
+            } finally {
+                System.out.println("\nGatewayAgent: Command Released\n");
+            }
+            System.out.println("\n=== GatewayAgent: Response sent back to controller ===\n");
+        }
+        else {
             System.out.println("Not working.");
         }
     }
+
+
+    @Override
+    protected void setup() {
+        super.setup();
+        System.out.println("\nGatewayAgent - setup - Agent " + getAID().getName() + " is ready.\n");
+        addBehaviour(new CyclicBehaviour() {
+            @SneakyThrows
+            @Override
+            public void action() {
+                System.out.println("\n=== GatewayAgent: Receiving Msg ====\n");
+                ACLMessage msg = receive();
+                if (msg != null) {
+                    System.out.println("\n=== GatewayAgent: Msg Rcd ====\n"+msg);
+                    // Process incoming messages
+                    String content = msg.getContent();
+                    System.out.println("BackGW: "+ content);
+                    JSONObject json = new JSONObject(content);
+                    String correlationId = json.getString("correlationId");
+
+                    AgentCommand command = pendingCommands.get(correlationId);
+
+                    if (command != null) {
+                        command.completeFutureResult(content);
+                        pendingCommands.remove(correlationId);
+                        RestaurantGatewayAgent.this.releaseCommand(command);
+                    } else {
+                        System.out.println("No matching command found for correlationId: " + correlationId);
+                    }
+
+                } else {
+                    System.out.println("\nMessage is null: "+ msg +"\n");
+                    block();
+                }
+            }
+        });
+    }
+
+
+
 }
